fix: Button as separate styled component
